Modified the code based on PR comments of TicTacToe
