+++ src/main/java/org/example/README.md
@@ -1,0 +1,18 @@
+FlashCards
+
+Instructions:
+1. Run the program from FlashCardArchive
+2. From the menu, select the desired action. At first try to add flashcards in order to be able to use the program properly
+3. A flashcard is a digital or paper card that contains a term on one side, and a definition, translation, or explanation 
+     of that term on the other. Flashcards are often used for learning foreign languages and are an effective study technique for many people.
+4. For adding FlashCard, you need to enter the name of the Card and its definition.
+5. After adding several cards, you can try yourself by choosing ask action. It will ask you for definitions of those cards that you have added.
+6. It keeps track of your answers, and if you answer wrong, it stores that data, so when you choose the hardest card action, 
+     it will give you the card on which you have most errors, you can reset those errors by "reset stats"
+7. The Program keeps the log too, so you can export it by "log" action.
+8. You can also input several flashCards from a file by import action, or by giving a terminal argument like this:
+   "java Flashcards -import yourFile.txt"
+9. You can also give the export folder in the terminal arguments like this:
+   "java Flashcards -import yourFile1.txt -export yourFile2.txt" It will import flashCards at the start of the program, 
+   and export what's left at before exiting the program.
+10. Or you can give both
